requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
addict=2.4.0=pypi_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py39h06a4308_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pypi_0
bitarray=2.6.0=pypi_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
blosc=1.21.0=h4ff587b_1
braceexpand=0.1.7=pypi_0
brewer2mpl=1.4.1=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
brunsli=0.1=h2531618_0
bs4=0.0.1=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
catalogue=2.0.7=py39h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h74dc2b5_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
clip=1.0=pypi_0
clip-video-encode=1.3.0=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.5=py39h06a4308_0
coloredlogs=15.0.1=py39h06a4308_0
commonmark=0.9.1=pypi_0
contourpy=1.0.5=pypi_0
cryptography=37.0.1=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cymem=2.0.6=py39h295c915_0
cython=0.29.30=py39h6a678d5_0
cython-blis=0.7.7=py39hce1f21e_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2022.7.0=py39h06a4308_0
dask-glm=0.2.0=pypi_0
dask-ml=2023.3.24=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
distributed=2022.7.0=pypi_0
easydict=1.9=pypi_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
fairseq=0.12.2=pypi_0
ffmpeg=1.4=pypi_0
ffmpeg-python=0.2.0=pypi_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=pypi_0
fire=0.4.0=pypi_0
fonttools=4.37.2=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py39h7f8727e_0
fsspec=2022.1.0=pypi_0
ftfy=6.1.1=pypi_0
future=0.18.2=pypi_0
futures=3.0.5=pypi_0
fvcore=0.1.5.post20220512=pypi_0
gensim=4.1.2=py39h295c915_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
goslate=1.5.4=pypi_0
grpcio=1.42.0=py39hce63b2e_0
h5py=3.7.0=py39h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.11.1=pypi_0
humanfriendly=10.0=py39h06a4308_0
hydra-core=1.0.7=pypi_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py39hf0132c2_1
imageio=2.19.3=py39h06a4308_0
importlib-metadata=4.11.3=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
ipykernel=6.15.2=py39h06a4308_0
ipython=8.4.0=py39h06a4308_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jupyter_client=7.3.5=py39h06a4308_0
jupyter_core=4.10.0=py39h06a4308_0
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=pypi_0
kmeans-pytorch=0.3=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pyhd3eb1b0_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=7.84.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libzopfli=1.0.3=he6710b0_0
llvmlite=0.39.1=pypi_0
locket=1.0.0=py39h06a4308_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.6.0=pypi_0
matplotlib-inline=0.1.6=py39h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mmcv-full=1.7.0=pypi_0
mmdet=2.25.3=pypi_0
model-index=0.1.11=pypi_0
msgpack=1.0.4=pypi_0
multidict=6.0.2=py39h5eee18b_0
multipledispatch=0.6.0=pypi_0
murmurhash=1.0.7=py39h295c915_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_0
nltk=3.7=pyhd3eb1b0_0
numba=0.56.4=pypi_0
numpy=1.21.5=py39h6c91a56_3
numpy-base=1.21.5=py39ha15fc14_3
oauthlib=3.2.1=py39h06a4308_0
omegaconf=2.0.6=pypi_0
open-clip-torch=2.9.3=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openmim=0.3.2=pypi_0
openssl=1.1.1s=h7f8727e_0
ordered-set=4.1.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=pypi_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_1
pathy=0.6.1=py39h06a4308_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39hace64e9_1
pip=22.1.2=py39h06a4308_0
portalocker=2.5.1=pypi_0
prefetch-generator=1.0.1=pypi_0
preshed=3.0.6=py39h295c915_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=pypi_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.5=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=py39h5eee18b_0
pydictionary=2.0.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-cluster=1.6.0=py39_torch_1.12.0_cu113
pytorch-fid=0.3.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.2.1=pypi_0
pywavelets=1.3.0=py39h7f8727e_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py39h6a678d5_0
readline=8.1.2=h7f8727e_1
regex=2022.7.9=py39h5eee18b_0
requests=2.28.1=py39h06a4308_0
requests-oauthlib=1.3.0=py_0
rich=12.6.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
sacrebleu=2.2.0=pypi_0
scenedetect=0.6.1=pypi_0
scikit-image=0.19.2=py39h51133e4_0
scikit-learn=1.1.2=pypi_0
scipy=1.7.3=py39h6c91a56_2
seaborn=0.12.0=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=63.4.1=py39h06a4308_0
shellingham=1.3.1=pyhd3eb1b0_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py39h06a4308_0
snappy=1.1.9=h295c915_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=3.3.1=py39h79cecc1_0
spacy-legacy=3.0.9=py39h06a4308_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
spacy-model-en_core_web_sm=3.3.0=pyhd3eb1b0_0
sqlite=3.39.2=h5082296_0
srsly=2.4.3=py39h295c915_0
stack_data=0.2.0=pyhd3eb1b0_0
tabulate=0.8.10=pypi_0
tblib=1.7.0=pypi_0
tensorboard=2.9.0=py39h06a4308_0
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
tensorboardx=2.5.1=pypi_0
termcolor=2.0.1=pypi_0
terminaltables=3.1.10=pypi_0
thinc=8.0.15=py39hae6d005_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h1ccaba5_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.12.1=py39_cu113
torchtext=0.13.1=pypi_0
torchvision=0.13.1=py39_cu113
tornado=6.1=pypi_0
tqdm=4.64.0=py39h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
typer=0.4.1=py39h06a4308_0
typing-extensions=4.3.0=py39h06a4308_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.11=py39h06a4308_0
video2numpy=2.3.0=pypi_0
wasabi=0.9.1=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webdataset=0.1.103=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yapf=0.32.0=pypi_0
yarl=1.8.1=py39h5eee18b_0
youtube-dl=2021.12.17=pypi_0
zeromq=4.3.4=h2531618_0
zfp=0.5.5=h295c915_6
zict=3.0.0=pypi_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
